move dedup code to etl package
