--- cmd/dedup/dedup_integration_test.go
@@ -1,3 +1,0 @@
-// +build integration
-
-package main_test
